requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
google-auth==2.34.0
google-auth-transport-requests==2.0.0
